scripts/get_git: Add support for using GitHub tokens
You can define ROCKNIX_GHTOKEN in your ~/.rocknix/options file.
This is usefull if you are testing some development changes in a
private GitHub repo before pushing it to a public one.
This Token is sent in an additional Header only to GitHub.
Also this is restricted to packages that define PKG_USETOKEN=yes.